app/requirements.txt
ane-transformers==0.1.3
attrs==23.1.0
boto3==1.28.83
botocore==1.31.83
cattrs==23.1.2
certifi==2023.7.22
charset-normalizer==3.3.2
coremltools==7.1
exceptiongroup==1.1.3
filelock==3.13.1
fsspec==2023.10.0
huggingface-hub==0.17.3
idna==3.4
jmespath==1.0.1
llvmlite==0.41.1
loguru==0.7.2
more-itertools==10.1.0
mpmath==1.3.0
numba==0.58.1
numpy==1.26.1
openai-whisper==20231106
packaging==23.2
protobuf==3.20.1
pyaml==23.9.7
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
s3transfer==0.7.0
safetensors==0.4.0
six==1.16.0
sympy==1.12
tiktoken==0.5.1
tokenizers==0.14.1
torch==1.11.0
tqdm==4.66.1
transformers==4.35.0
typing_extensions==4.8.0
urllib3==2.0.7
